move utils directory
