+++ app/polls/onboarding/components/steps/InsightsStep.js
@@ -3,8 +3,8 @@ import { LuUsersRound } from 'react-icons/lu'
 import { TextInsight } from '../TextInsight'
 import { NumberInsight } from '../NumberInsight'
 import { DataVisualizationInsight } from '../DataVisualizationInsight'
-import { useContactsStats } from 'app/polls/hooks/useContactsStats'
-import { mapContactsStatsToCharts } from 'app/polls/utils/mapContactsStatsToCharts'
+import { useContactsStats } from '../../../hooks/useContactsStats'
+import { mapContactsStatsToCharts } from '../../utils/mapContactsStatsToCharts'
 
 export default function InsightsStep({ }) {
