+++ app/polls/onboarding/utils/mapContactsStatsToCharts.js
